Last minute design fix: Made usecases subscribe themselves: moved subscription to the usecase
