+++ antifraud_project/src/adapter/input/messaging/message_consumer.controller.ts
@@ -2,8 +2,6 @@ import { Controller, OnModuleDestroy, OnModuleInit } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { Logger } from '@nestjs/common';
 import { MessagingService } from '../../input_output/messaging/messaging.service';
-import { AntifraudCheckPayload } from '../../../domain/models/events/antifraud_check.payload';
-import { FraudAnalysisUsecase } from '../../../domain/usecases/fraud_analysis.usecase';
 
 @Controller()
 export class MessageConsumerController
@@ -11,7 +9,6 @@ export class MessageConsumerController
 {
   constructor(
     private readonly configService: ConfigService,
-    private readonly fraudAnalysisUsecase: FraudAnalysisUsecase,
     private readonly messagingService: MessagingService,
   ) {}
 
@@ -23,27 +20,6 @@ export class MessageConsumerController
 
     // Consumers
     Logger.log('MessageConsumerController - consumers');
-
-    // Consumer for topic "antifraud-check"
-    const antifraudCheckTopic = this.configService.get(
-      'application.transport.event-driven.kafka.topics.antifraud-check',
-    );
-
-    this.messagingService.addTopicConsumer(antifraudCheckTopic, (msg) => {
-      const checkPayload: AntifraudCheckPayload = JSON.parse(
-        msg.value.toString(),
-      );
-
-      Logger.log(
-        `>> ANTIFRAUD AntifraudConsumerController: read incoming message ` +
-          `${JSON.stringify(checkPayload)}`,
-      );
-
-      const transactionId: string = checkPayload.transactionId;
-      this.fraudAnalysisUsecase.analyze(transactionId);
-    });
-
-    this.messagingService.initializeConsumers();
   }
 
   async onModuleDestroy() {
