+++ transaction_project/src/domain/usecases/update_transaction_after_validation.usecase.ts
@@ -1,11 +1,41 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, OnModuleInit } from '@nestjs/common';
 import { Logger } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
+import { MessagingService } from 'src/adapter/input_output/messaging/messaging.service';
 import { TransactionService } from '../../adapter/output/db/transaction.service';
 import { AntifraudAnalysisResponsePayload } from '../models/events/antifraud_analysis_response.payload';
 
 @Injectable()
-export class UpdateTransactionAfterValidationUsecase {
-  constructor(private transactionService: TransactionService) {}
+export class UpdateTransactionAfterValidationUsecase implements OnModuleInit {
+  onModuleInit() {
+    // Consumer for topic "antifraud-analysis-response"
+    const antifraudAnalysisResponseTopic = this.configService.get(
+      'application.transport.event-driven.kafka.topics.antifraud-analysis-response',
+    );
+
+    this.messagingService.addTopicConsumer(
+      antifraudAnalysisResponseTopic,
+      (msg) => {
+        const analysisResponse: AntifraudAnalysisResponsePayload = JSON.parse(
+          msg.value.toString(),
+        );
+        Logger.log(
+          `UpdateTransactionAfterValidationUsecase antifraud-analysis-response ` +
+            `payload.transactionId : ${analysisResponse.transactionId}`,
+        );
+
+        this.update(analysisResponse);
+      },
+    );
+
+    this.messagingService.subscribeConsumers();
+  }
+
+  constructor(
+    private messagingService: MessagingService,
+    private transactionService: TransactionService,
+    private configService: ConfigService,
+  ) {}
 
   update(antifrauAnalysisResponse: AntifraudAnalysisResponsePayload) {
     Logger.log(
